+++ composer.json
@@ -9,7 +9,7 @@
         }
     ],
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.1.3",
         "comoyi/redis": "^2.1",
         "symfony/console": "^3.3",
         "illuminate/config": "^5.4",
